pyproject.toml
[project]
name = "dc_crime"
version = "0.1.0"
description = "Data journalism project: dc_crime"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "censusdis>=1.4.0",
    "folium>=0.15.0",
    "geopandas>=1.0.0",
    "ipywidgets>=8.1.7",
    "jupyter>=1.1.1",
    "jupyterlab>=4.4.5",
    "lxml>=5.3.0",
    "matplotlib>=3.10.5",
    "nbconvert>=7.16.4",
    "notebook>=7.4.5",
    "numpy>=2.3.2",
    "openpyxl>=3.1.0",
    "pandas>=2.3.1",
    "plotly>=6.2.0",
    "polars>=1.32.2,<1.33.0",
    "psycopg2-binary>=2.9.0",
    "pyarrow>=18.0.0,<19.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "requests>=2.32.4",
    "scikit-learn>=1.7.1",
    "seaborn>=0.13.2",
    "selenium>=4.25.0",
    "sqlalchemy>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "utils"]
